requirements.txt
flask==2.0.2
nba_api
gunicorn
python-dateutil
beautifulsoup4